+++ src/EventManager.cpp
@@ -13,21 +13,18 @@
 EventManager::EventManager(){
   event_queue a;
   event_queue b;
-  queues.push_back(a);
-  queues.push_back(b);
-  process_queue = &queues.at(0);
-  register_queue = &queues.at(1);
+  queues.push_back(make_shared<event_queue>(a));
+  queues.push_back(make_shared<event_queue>(b));
+  process_queue = queues.at(0);
+  register_queue = queues.at(1);
 }
 
 EventManager::~EventManager(){
-  process_queue = nullptr;
-  register_queue = nullptr;
+
 }
 
 //add an event to the queue of registered events
 void EventManager::queueEvent(shared_ptr<EventInterface> event){
-  cout << register_queue -> size() << endl;
-  cout << process_queue -> size() << endl;
   //in case we do not find the event type we add it
   //to the map
   event_map::const_iterator it = EventDelegateMap.find(event -> getEventType());
@@ -41,8 +38,6 @@ void EventManager::queueEvent(shared_ptr<EventInterface> event){
 
 //add an event to the map of events and their delegates
 void EventManager::registerEvent(const EventType &type){
-  cout << register_queue -> size() << endl;
-  cout << process_queue -> size() << endl;
   event_delegates newdelegateList;
   event_map::const_iterator it = EventDelegateMap.find(type);
 
@@ -55,8 +50,6 @@ void EventManager::registerEvent(const EventType &type){
 //register a delegate to an event type. Think of this like a person (class) is subscribing
 // to a particular magazine (EvenetType) and is sending their address (EventDelegate)
 void EventManager::registerDelegate(const EventDelegate& d, const string &eventDelegateIdentifier,const EventType &type){
-  cout << register_queue -> size() << endl;
-  cout << process_queue -> size() << endl;
   //in case we do not find the event type we add it
   //to the map
   event_map::const_iterator it = EventDelegateMap.find(type);
@@ -74,8 +67,6 @@ void EventManager::registerDelegate(const EventDelegate& d, const string &eventD
 }
 //deregister a delegate to an event type
 void EventManager::deregisterDelegate(const EventDelegate& d, const string &eventDelegateIdentifier, const EventType& type){
-  cout << register_queue -> size() << endl;
-  cout << process_queue -> size() << endl;
   //look for the delegate to see if it is already in the list
   event_delegates eventDelegatesEventMap = EventDelegateMap.at(type);
   if(!(eventDelegatesEventMap.find(eventDelegateIdentifier) == eventDelegatesEventMap.end())){
@@ -87,8 +78,6 @@ void EventManager::deregisterDelegate(const EventDelegate& d, const string &even
 
 void EventManager::triggerEvent(const EventInterface& event)
 {
-  cout << register_queue -> size() << endl;
-  cout << process_queue -> size() << endl;
   //search to see if this event is registered
   auto delegateList = EventDelegateMap.find(event.getEventType());
   if(delegateList == EventDelegateMap.end()){
@@ -108,8 +97,6 @@ void EventManager::triggerEvent(const EventInterface& event)
 
 //iterate through all registered events and trigger them
 void EventManager::processEvent(){
-  cout << register_queue -> size() << endl;
-  cout << process_queue -> size() << endl;
   std::swap(process_queue, register_queue);
   clear(*register_queue);
   while(!process_queue->empty()){
@@ -143,5 +130,6 @@ void EventManager::clear( event_queue &q )
    if(q.size() != 0){
      event_queue empty;
      std::swap( q, empty );
+
    }
 }